Better utilize security packet.
I've figued out that the security data packet can be used to store
information between sessions since it's just echoed back to the server.
The other servers use this to figure out when to send the client along
to the shipgate, so I've made adjustments to be able to do the same.