Add base logic for pulling ship list from shipgate
The shipgate now spins up an Http server that handles requests
from a client for details about the available ships. This is so
that the login server doesn't need to manage any of the ship
connection details and the ship list can be determined on the
fly, propogating from the ship server itself to login.