Fixed erratic behavior when no patch dir exists
Old code didn't check to see if a move into the patch directory was successful, using any files in
the current directory for the patch list.
New code not only tests if directory exists, but makes it (if it can) if it doesn't, and additionally
exits if it can't enter newly made patch directory.