Added external facing IP configuration
Previously when the server would send redirect addresses to the client, it would send the
IP address the server is bound to.  However, for servers behind routers, the external-facing
address is not the same as the bind address, causing the client to fail to reconnect after a
redirect is sent.

Now the external IP address is declarable in the config json.