Remove the packet analyzer
It was too confusing for new people to use and limited in that it
needed to be wired in to the server you were working with. I've
replaced it with a client-side packet sniffer that will make it
possible to sniff other servers (someday, hopefully) and avoid
needing to configure/implement the connection points to the analyzer.