Select the correct character in the block server
I haven't been sure how to determine which slot the client had
chosen since it didn't seem to appear anywhere. Tethealla sticks
this in its security information after receiving the character
selection packet in the CHAR server, but in trying to avoid that I
realized that a field in the login packet actually has the chosen
slot in there. So let's use that instead.